Move all topical routes to /actueel
